First sync of Research
Adding Geo based map.

In 2013, an anonymous hacker mapped the Internet through illegal means,
and in the process exposed rampant security problems. The project,
called Internet Census 2012, used 420,000 networked devices, dubbed the
Carna Botnet, to ping IP addresses across the globe in 2012. Every one
of the devices was either entirely unsecured with no password
protection, or used the standard password "root" that comes with many
off-the-shelf routers (users are supposed to change the password, but
rarely do). The hacker released all of the collected data to the public
domain in a sort of research paper. The animation above is a map based
on that data that shows 24-hours of Internet use.