Logical map
This topological map of the Internet from 2007 is the result of years
of research into the structure and connections of the Internet. The
core of the structure contains around 80 nodes that have a lot of
connections with other well-connected nodes. surrounding the core is a
region of around 5,000 less-connected nodes that are dependent on the
core. Beyond that lies a ring of around 15,000 more self-sufficient,
peer-connected nodes.