2000 era relationship
This is a topological map of the Internet from May 30, 2015 created by
The Internet Mapping Project, which was one of the first major mapping
efforts. It was started in 1998 at Bell Labs and then moved to a
Lucent/Bell Labs spin-off called Lumeta in 2000. The lines on this map
represent the shortest outgoing route from a test computer at Lumeta in
Somerset, New Jersey to 450,000 registered nets on the Internet. The
creators of this map used programs that arrange the map based on rules
that simulate springs and repelling forces. See a close-up on the next
slide. LUMETA
Gallery Image